+++ src/Catalog/Declination.php
@@ -50,6 +50,9 @@ final class Declination
     /** @var Image[] */
     private $images;
 
+    /** @var bool */
+    private $isBrandNew;
+
     /** @var DeclinationOption[] */
     private $options;
 
@@ -77,6 +80,7 @@ final class Declination
         $this->images = array_map(static function (array $imageData) : Image {
             return new Image($imageData);
         }, $data['images']);
+        $this->isBrandNew = $data['isBrandNew'] ?? true;
         $this->options = array_map(static function (array $optionData) : DeclinationOption {
             return new DeclinationOption($optionData);
         }, $data['options']);
@@ -141,6 +145,14 @@ final class Declination
     public function getAffiliateLink(): ?string
     {
         return $this->affiliateLink;
+    }
+
+    /**
+     * Is the declination brand new (true) or second-hand (false).
+     */
+    public function isBrandNew(): bool
+    {
+        return $this->isBrandNew;
     }
 
     /**